+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Safely handle an empty values file being provided by replacing it with a single newline character
 
+### Changed
+
+- Updated `.WithVersion()` to better handle sha based versions and set the catalog to the currently set catalog with a `-test` suffix (instead of hardcoded to `cluster-test`)
+
 ## [0.14.0] - 2023-12-04
 
 ### Changed